Fix typo

+++ include/json_parse_utils.h
@@ -157,7 +157,7 @@ struct JsonFormatError
 {                                                                              \
     JSON_ASSERT_CONTAINS((NODE), (ATTR_NAME));                                 \
     JSON_ASSERT_INT_64((NODE)[(ATTR_NAME)]);                                   \
-    (TARGET) = (NODE)[(ATTR_NAME)].GetInt();                                   \
+    (TARGET) = (NODE)[(ATTR_NAME)].GetInt64();                                 \
 }
 
 #define JSON_GET_BOOL_MEMBER(NODE, ATTR_NAME, TARGET)                          \
